Updates for newest MP Telemetry TCK tests
