+++ appserver/payara-appserver-modules/microprofile/telemetry/src/main/java/fish/payara/microprofile/telemetry/tracing/jaxrs/OpenTracingCdiUtils.java
@@ -41,25 +41,19 @@
  */
 package fish.payara.microprofile.telemetry.tracing.jaxrs;
 
+import fish.payara.microprofile.telemetry.tracing.WithSpanMethodInterceptor;
 import jakarta.enterprise.inject.spi.BeanManager;
 import jakarta.interceptor.InvocationContext;
 import jakarta.ws.rs.container.ResourceInfo;
 import org.eclipse.microprofile.config.Config;
 import org.eclipse.microprofile.config.ConfigProvider;
-import org.jboss.weld.interceptor.WeldInvocationContext;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
-import java.util.LinkedList;
-import java.util.Objects;
 import java.util.Optional;
-import java.util.Queue;
-import java.util.Set;
-import java.util.function.Function;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import static java.util.Arrays.asList;
 import static java.util.logging.Level.FINE;
 import static java.util.logging.Level.FINER;
 import static java.util.logging.Level.FINEST;
@@ -80,74 +74,14 @@ final class OpenTracingCdiUtils {
      * @param <A> The annotation type to return
      * @param beanManager The invoking interceptor's BeanManager
      * @param annotationClass The class of the annotation to get
-     * @param invocationContext The context of the method invocation
-     * @return The annotation that triggered the interceptor.
-     */
-    public static <A extends Annotation> A getAnnotation(BeanManager beanManager, Class<A> annotationClass, InvocationContext invocationContext) {
-        A annotation = getInterceptedAnnotation(annotationClass, invocationContext);
-
-        if (annotation == null) {
-            annotation = getAnnotation(beanManager, annotationClass, invocationContext.getMethod().getDeclaringClass(), invocationContext.getMethod());
-        }
-
-        return annotation;
-    }
-
-    /**
-     * Gets the annotation from the method that triggered the interceptor.
-     *
-     * @param <A> The annotation type to return
-     * @param beanManager The invoking interceptor's BeanManager
-     * @param annotationClass The class of the annotation to get
      * @param resourceInfo The targeted jaxrs resource
      * @return The annotation that triggered the interceptor.
      */
     public static <A extends Annotation> A getAnnotation(BeanManager beanManager, Class<A> annotationClass, ResourceInfo resourceInfo) {
-        return getAnnotation(
+        return WithSpanMethodInterceptor.getAnnotation(
             beanManager, annotationClass,
             resourceInfo.getResourceClass(),
             resourceInfo.getResourceMethod());
-    }
-
-    public static <A extends Annotation> A getAnnotation(BeanManager beanManager, Class<A> annotationClass, Class<?> annotatedClass, Method method) {
-        logGetAnnotation(annotationClass, method);
-        Objects.requireNonNull(annotatedClass, "annotatedClass");
-        Objects.requireNonNull(method, "method");
-
-        // Try to get the annotation from the method, otherwise attempt to get it from the class
-        if (method.isAnnotationPresent(annotationClass)) {
-            LOG.log(FINER, "Annotation was directly present on the method");
-            return method.getAnnotation(annotationClass);
-        }
-
-        if (annotatedClass.isAnnotationPresent(annotationClass)) {
-            LOG.log(FINER, "Annotation was directly present on the class");
-            return annotatedClass.getAnnotation(annotationClass);
-        }
-
-        LOG.log(FINER, "Annotation wasn't directly present on the method or class, checking stereotypes");
-
-        // Account for Stereotypes
-        Queue<Annotation> annotations = new LinkedList<>(asList(annotatedClass.getAnnotations()));
-
-        // Loop over each individual annotation
-        while (!annotations.isEmpty()) {
-            Annotation a = annotations.remove();
-
-            // Check if this is the annotation we're looking for
-            if (a.annotationType().equals(annotationClass)) {
-                LOG.log(FINER, "Annotation was found in a stereotype");
-                return annotationClass.cast(a);
-            }
-
-            // If the found annotation is a stereotype, get the individual annotations and add them to the list
-            // to be iterated over
-            if (beanManager.isStereotype(a.annotationType())) {
-                annotations.addAll(beanManager.getStereotypeDefinition(a.annotationType()));
-            }
-        }
-
-        return null;
     }
 
     /**
@@ -272,25 +206,6 @@ final class OpenTracingCdiUtils {
         return appValue;
     }
 
-    /**
-     * Uses Weld to find an annotation. The annotation must be an intercepted annotation that has a binding
-     * @param <A>
-     * @param annotationClass {@link Annotation} to look for
-     * @param invocationContext
-     * @return the {@link Annotation} that the interceptor is inceptoring,or null if it cannot be found
-     */
-    private static <A extends Annotation> A getInterceptedAnnotation(Class<A> annotationClass, InvocationContext invocationContext){
-         if (invocationContext instanceof WeldInvocationContext) {
-            Set<Annotation> interceptorBindings = ((WeldInvocationContext) invocationContext).getInterceptorBindings();
-            for (Annotation annotationBound : interceptorBindings) {
-                if (annotationBound.annotationType().equals(annotationClass)) {
-                    return (A) annotationBound;
-                }
-            }
-        }
-        return null;
-    }
-
     private static Config getConfig() {
         try {
             return ConfigProvider.getConfig();
@@ -301,17 +216,5 @@ final class OpenTracingCdiUtils {
         return null;
     }
 
-    private static void logGetAnnotation(Class<?> annotatedClass, Method method) {
-        LOG.log(FINER, "Attempting to get annotation {0} from {1}",
-            new String[] {getString(annotatedClass, Class::getSimpleName), getString(method, Method::getName)});
-    }
-
-
-    private static <T> String getString(final T object, Function<T, String> toString) {
-        if (object == null) {
-            return null;
-        }
-        return toString.apply(object);
-    }
 
 }
