+++ nucleus/packager/external/opentelemetry-repackaged/pom.xml
@@ -72,6 +72,7 @@
         <dependency>
             <groupId>io.opentelemetry.instrumentation</groupId>
             <artifactId>opentelemetry-instrumentation-annotations</artifactId>
+            <version>${opentelemetry.version}</version>
             <scope>compile</scope>
             <optional>true</optional>
             <exclusions>
@@ -85,6 +86,7 @@
         <dependency>
             <groupId>io.opentelemetry.instrumentation</groupId>
             <artifactId>opentelemetry-instrumentation-annotations-support</artifactId>
+            <version>${opentelemetry.alpha.version}</version>
             <scope>compile</scope>
             <optional>true</optional>
             <exclusions>
